==> 4-Analysis_Corona.Rmd <==
---
title: "Analysing legislative behaviour before and after March 2020"
author: "Mio Hienstorfer-Heitmann"
date: "29 1 2021"
output:
  html_document:
    mathjax: local
    self_contained: false
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      include = TRUE,
                      warning = FALSE,
                      message = FALSE)
```

## Introduction

```{r packages, include =FALSE}

p_needed <- c("knitr","tidyverse","texreg","optimx","MASS","kableExtra","table1")
packages <- rownames(installed.packages())
p_to_install <- p_needed[!(p_needed %in% packages)]
if (length(p_to_install) > 0) {
  install.packages(p_to_install)
}
lapply(p_needed, require, character.only = TRUE)


```


```{r load and process data}



df <- read_csv2("Dataframe_opp_corona.csv")


# Only analyse opposition parties
df_opposition <- df %>% 
  filter(Opposition == 1) %>%
  filter(Fraktion %in% c("FDP","CDU/CSU","AfD","DIE GRÜNEN","DIE LINKE","SPD")) %>%
  mutate(Health = as.numeric(Health),
         Corona_leg = ifelse(grepl("[Cc]orona", Name),1,0),
         AfD = ifelse(Fraktion == "AfD",1,0))

logit_df <- df_opposition %>%
  dplyr::select("Party_yes", "Corona","Days_next_ele", "Rel_Seats", "Rile_Dist", "AfD") %>%
  mutate(Rile_DistCorona = Corona * Rile_Dist) %>%
  na.omit
X <- as.matrix(cbind(1,logit_df[,2:ncol(logit_df)]))

y <- logit_df[,1]

  
```

In the following, I will conduct a short analysis of Bundetags-votes by opposition parties in the 19th legislative period between 2018 and the autumn of 2020. 

```{r descriptive statistics}

# Number of votes
# Number of parties
# seat shares
# distance to gov
# What period is covered
# days until next election
#

df_opposition <- df_opposition %>%
  mutate(Corona = as.factor(ifelse(Corona == 1, "Post Corona", "Pre Corona")))

table1::label(df_opposition$Rel_Seats) <- "Relative Seat share"
table1::label(df_opposition$Rile_Dist) <- "Distance to Government"
table1::label(df_opposition$Days_next_ele) <- "Days until next election"
table1::label(df_opposition$Corona) <- "Share Votes after Corona"




table1::table1(~Rel_Seats + Rile_Dist + Days_next_ele | Fraktion+ Corona, data = df_opposition)




```


The above table shows descriptive statistics for all of the variables in my model. The main explanatory factor to explain differences in opposition voting behaviour is a dummy indicating the time before (0) and after (1) the onset of the Corona pandemic in Germany in march 2020. I further include the relative seat share of a party into the model, as well as its ideological distance to the SPD-CDU/CSU led coalition as measured by the CMP-data.
I further include the days until the next election in autumn of 2021, indicating that the closer an election, the more confronting opposition parties are likely to be in parliament. As already becomes appearent, the number of roll call votes in the Bundestag after March 2020 is fairly limited. We should thus not expect to observe significant differences in behaviour before and after Corona.

## Model choice

When analysing the effects of party- as well as vote-specific characteristics on an outcome variable, one promising approach would be to apply a multilevel model with random intercepts and/or slopes on the vote- as well as the party level. On the other hand, such models have the tendency to overfit, further they are not easily interpreted and may not converge with an N too small. That is why I will apply a logistic regression model as:

\[P(Vote = 1 | X = X_i ) = P(Vote = 1) = \frac{1}{1 + exp(-x_i\beta)}\]

whereas my systematic component $x_i\beta$ is:

\[X_i\beta = (x_1\beta_{Corona},x_2\beta_{DaysNextElection},x_3\beta_{LeftRightDistance},x_4\beta_{RelSeatShare},x_5\beta_{AfD})\]

I thus expect the likelihood to vote in line with the government coalition in Germany is a function of whether the vote happend before or after Corona, the distance to the next election, the distance of a party to the coalition in ideological terms and the relative seat share of a party. I further include the AfD as a control since it is a party with unsual legislative behaviour as a populist and I thus want to account for the possibility that the AfD may yield patterns of behaviour that are otherwise left unexplained by the model.

I apply a BFGS-hill-climbing algorithm to estimate the maximum likelihood of observing a positive vote by an opposition party given the systematic component $X_i\beta$.


## Results

The following table presents the results of the regression analysis. 

```{r logit model}

ll_logit <- function(theta, y, X) {
  # theta consists merely of beta (dim is ncol(X))
  beta <- theta[1:ncol(X)]
  # linear predictor; make sure that X is stored as.matrix
  mu <- X %*% beta
  # response function
  p <- 1 / (1 + exp(-mu))
  # log-likelihood
  ll <- y * log(p) + (1 - y) * log(1 - p)
  # sum
  ll <- sum(ll)
  return(ll)
}



# start values
startvals <- rep(0,ncol(X))


# optimize

res <- optim(
  par = startvals,
  fn = ll_logit,
  y = y,
  X = X,
  control = list(fnscale = -1),
  hessian = TRUE,
  method = "BFGS"
)

### Format the output

format_probit_output <- function(res, X, Caption) {

    if (is.null(colnames(X))) {
    var_names <- c("Intercept", paste("Var", 1:(ncol(X) - 1)))
    colnames(X) <- var_names
  }
  # Estimates:
  probit_coef <- round(res$par,3)
  # Standard Errors:
  probit_se <- round(sqrt(diag(solve(-res$hessian))),3)
  
  # t value:
  t <- round(probit_coef / probit_se,3)
  # degrees of freedom:
  df <- nrow(X) - length(probit_coef)
  t.val <- qt(0.975, df) # critical value of t
  # p value:
  p <-  round(2 * pt(-abs(t), df),3)
  
  out <- cbind(probit_coef, probit_se, p)[1:ncol(X), ]
  
  colnames(out) <- c(" ML Estimates ", " ML SE ", " Pr(>|t|) ")
  
  if(is.null(colnames(X[,1]))){
    
    colnames(X) <- c("Intercept",colnames(X[,2:ncol(X)]))

  }
  
  rownames(out) <- colnames(X)
  
  
  # Headline
  #cat(paste0("\n",Caption,"\n"), rep("=", 40), "\n", sep = "")
  #print(out)
  return(model = out)
}


kbl(format_probit_output(res = res, X = X, Caption = "Probit Model of Legislative Voting"))



```

First it is apparent that none of the coefficient except for the ideological difference of a party to the coalition yields significant effects. It is unsurprisingly shown that increasing distance to the government decreases the probability to vote with the government in a roll call vote. Regarding the Corona-pandemic, we observe a positive effect for our sample that remains insignificant.

### Simulation

In the following I will simulate the effect of the Corona pandemic on legislative behaviour. I will further examine whether that behaviour was different for parties with different ideological positions. That is: did parties behave differently after Corona depending on their ideological distance to the coalition?

```{r simulation}

### ========================= Simulation ### ===================##

### Sim function


sim_function <-
  function(nsim = 1000,
           coefs,
           vcov,
           scenario,
           response_function,
           predicted_values = F,
           stochastic_component) {
    if (is.null(dim(scenario))) {
      stop("The scenario needs to be in a matrix.")
    }
    if (length(coefs) != ncol(scenario)) {
      stop("The scenario and the parameter vector don't fit.")
    }
    # Set up the sampling distribution
    S <- mvrnorm(nsim, coefs, vcov)
    
    # Calculate mu and p
    
    # The linear component
    mu <- S %*% t(scenario)
    
    # The response function
    p <- response_function(mu)
    
    # Calculate ev, observed value approach
    ev <- apply(p, 1, mean) 
    
    if (predicted_values) {
      pv <-
        array(stochastic_component(ndraws = prod(dim(ev)), p = ev), dim = dim(ev))
      return(list(ev = ev, pv = pv))
    }
    return(list(ev = ev))
    
  }


### calculate:how did Corona crisis affect opposition work? Also: 

# logit response function
response_function <- function(x) {
  1 / (1 + exp(-x))
}

# for predicted values: the stochastic component
stochastic_component <- function(ndraws, p) {
  rbinom(n = ndraws, size = 1, prob = p)
}

# coefficients and variance-covariance matrix of our model
coefs <- res$par[1:ncol(X)]
vcov <- solve(-res$hessian[(1:ncol(X)), (1:ncol(X))])

# Which scenarios to simulate for holding all else equal

scenarios_corona <- c(0,1)
scenarios_distance <- sort(unique(df_opposition$Rile_Dist))


cases <- array(NA, c(dim(X), # 1 & 2
                     length(scenarios_corona), # 3
                     length(scenarios_distance))) # 4

cases[, , ,] <- X

# scenarios
for(i in 1:length(scenarios_corona)){
  for(j in 1:length(scenarios_distance)){
    cases[,which(colnames(X) == "Corona"),i,] <- scenarios_corona[i]
    cases[,which(colnames(X) == "Rile_Dist"),,j] <- scenarios_distance[j]
    
    
  }
}

# interaction between scenarios

for(i in 1:length(scenarios_corona)){
  for(j in 1:length(scenarios_distance)){
  cases[,which(colnames(X) == "Rile_DistCorona"),i,] <- cases[,which(colnames(X) == "Rile_Dist"),i,] * scenarios_corona[i]
  cases[,which(colnames(X) == "Rile_DistCorona"),,j] <- cases[,which(colnames(X) == "Corona"),,j] * scenarios_distance[j]
  }
}


ev_corona <- matrix(NA, nrow = 1000, ncol = length(scenarios_corona)*length(scenarios_distance))

for(i in 1:length(scenarios_corona)){
  for(j in 1:length(scenarios_distance)) {
    if(i == 1){
      ev_corona[,i*j] <- sim_function(
        coefs = coefs,
        vcov = vcov,
        response_function = response_function,
        stochastic_component = stochastic_component,
        scenario = cases[,,i,j],
        predicted_values = F
      )$ev
    } else{
        ev_corona[,i+2+j] <- sim_function(
          coefs = coefs,
          vcov = vcov,
          response_function = response_function,
          stochastic_component = stochastic_component,
          scenario = cases[,,i,j],
          predicted_values = F
        )$ev
      }
  }
}



ev_corona_mean <- apply(ev_corona, 2, mean)
ev_corona_ci <- t(apply(ev_corona, 2, quantile, c(0.025, 0.975)))


df_corona <- data.frame(
  mean = ev_corona_mean,
  lower = ev_corona_ci[,1],
  upper = ev_corona_ci[,2],
  scenario = c(paste0("Pre Corona_",sort(unique(df_opposition$Rile_Dist))),
               paste0("Post Corona_",sort(unique(df_opposition$Rile_Dist))))
)

df_corona <- df_corona %>%
  separate(scenario, sep = "_", into = c("Corona","Rile_Dist")) %>%
  mutate(Rile_Dist = as.numeric(Rile_Dist)
         # Party = case_when(max(Rile_Dist) ~ "The Left",
         #                   min(Rile_Dist) ~ "FDP",
         #                   sort(Rile_Dist)[2] ~ "The Greens",
         #                   sort(Rile_Dist)[3] ~ "AfD")
         )

ggplot(data = df_corona, aes(x = Rile_Dist, color = Corona))+
  geom_line(aes(y = mean))+
  geom_line(aes(y = lower), linetype = "dashed")+
  geom_line(aes(y = upper), linetype = "dashed")+
  ylab("P(Voting with Gov.)")+
  xlab("Left-Right Distance to Coalition")+
  theme_bw()


```




### Limitations

It would make more sense to write a multilevel model to account for variations between the party- and the vote level. Maybe effects are different across parties (would take away entire variation in seatshare and distance to gov). On the vote level, we would have variations in the days until the next election.
